Run decoupling detection in capacitor source render
