--- lib/utils/decoupling-capacitors/port-is-ground.ts
@@ -1,31 +1,0 @@
-import type { PinAttributeMap } from "@tscircuit/props"
-import type { Port } from "lib/components/primitive-components/Port"
-import { GROUND_NET_REGEX } from "lib/utils/gnd-power-net-regex"
-
-/** Resolves whether a port represents ground from its attributes or labels. */
-export function portIsGround(sourcePort: Port): boolean {
-  const sourceComponent = sourcePort.getParentNormalComponent()
-  const pinAttributes = sourceComponent?._parsedProps.pinAttributes as
-    | Record<string, PinAttributeMap>
-    | undefined
-  let requiresGround: boolean | undefined
-  let providesGround: boolean | undefined
-
-  for (const portName of sourcePort.getNameAndAliases()) {
-    const portAttributes = pinAttributes?.[portName]
-    if (!portAttributes) continue
-    if (portAttributes.requiresGround !== undefined) {
-      requiresGround = portAttributes.requiresGround
-    }
-    if (portAttributes.providesGround !== undefined) {
-      providesGround = portAttributes.providesGround
-    }
-  }
-
-  if (requiresGround === true || providesGround === true) return true
-
-  for (const portName of sourcePort.getNameAndAliases()) {
-    if (GROUND_NET_REGEX.test(portName)) return true
-  }
-  return false
-}
